NOTE - CLEAN BUILD REQUIRED! ASH ints are now Java longs. ASH floats are now Java doubles. ASH booleans are now Java BigBoo... no wait, they were fine as they are. There is no increase in runtime memory requirements; since no Value needs both int and float content, they can be both stored in the same field. Known limitations at this point: * There still isn't a way to get substat or meat amounts beyond 2 billion, but at least that's now a possibility. * Ints outside the 32-bit range, and floats outside the single-precision range, probably cannot be stored in a map file at the moment, nor survive round-trip conversion to strings. * Ints outside the 32-bit range, when passed to most library functions, will simply have the top 32 bits ignored, rather than doing any range checking on the full 64-bit value (so 2**32+7 is exactly the same as 7 in most cases).
git-svn-id: http://svn.code.sf.net/p/kolmafia/code@11059 b29ace70-8910-0410-8dcc-aa2fc6433167

+++ src/net/sourceforge/kolmafia/textui/parsetree/ArrayValue.java
@@ -57,7 +57,7 @@ public class ArrayValue
 	public Value aref( final Value key, final Interpreter interpreter )
 	{
 		Value[] array = (Value[]) this.content;
-		int index = key.intValue();
+		int index = (int) key.intValue();
 		if ( index < 0 || index >= array.length )
 		{
 			throw interpreter.runtimeException( "Array index out of bounds" );
@@ -69,7 +69,7 @@ public class ArrayValue
 	public void aset( final Value key, final Value val, final Interpreter interpreter )
 	{
 		Value[] array = (Value[]) this.content;
-		int index = key.intValue();
+		int index = (int) key.intValue();
 		if ( index < 0 || index >= array.length )
 		{
 			throw interpreter.runtimeException( "Array index out of bounds" );
@@ -104,7 +104,7 @@ public class ArrayValue
 	public Value remove( final Value key, final Interpreter interpreter )
 	{
 		Value[] array = (Value[]) this.content;
-		int index = key.intValue();
+		int index = (int) key.intValue();
 		if ( index < 0 || index >= array.length )
 		{
 			throw interpreter.runtimeException( "Array index out of bounds" );
@@ -135,7 +135,7 @@ public class ArrayValue
 	public boolean contains( final Value key )
 	{
 		Value[] array = (Value[]) this.content;
-		int index = key.intValue();
+		int index = (int) key.intValue();
 		return index >= 0 && index < array.length;
 	}
